requirements.txt
annotated-types==0.7.0
bcrypt==3.1.7
fastapi==0.115.11
passlib==1.7.4
PyJWT==2.10.1
python-dotenv==1.0.1
uvicorn==0.34.0
streamlit==1.43.2
